# Add Merkle-optimized BLAKE2b implementation