… merkle/blake2b/blake2b_amd64.s
… merkle/merkle_test.go