I wrote custom BLAKE2b assembly for working with Sia Merkle trees. The assembly requires AVX2, so not everyone will be able to take advantage of it. I figure writing a separate AVX version is probably not worth at this point, but I would like to write an AVX-512 version eventually. Other archs would be good too, especially ARM, but the Go toolchain doesn't support the NEON instruction set yet. :( Also I don't have an ARM machine to test with...

The `merkle` package is a bit more complex now, but it's not as horrible as I feared, even though I probably went a little overboard. :P

Anyway, here are the new download benchmarks:

|                | 100 Mbps | 500 Mbps |  1 Gbps  |  2 Gbps  |  6 Gbps   |
|----------------|----------|----------|----------|----------|-----------|
|  **Before**    |  64 Mbps | 192 Mbps | 320 Mbps | 440 Mbps |  800 Mbps |
|  **After**     |  80 Mbps | 256 Mbps | 480 Mbps | 660 Mbps | 1720 Mbps |
|  Increase      |    25%   |   33%    |   50%    |   150%   |   215%    |

As expected, the more bandwidth you have, the more you benefit from this optimization. But even at 100 Mbps, 25% isn't bad!

Proof construction also got much faster, although you won't see the benefits outside of the `us` test suite -- mainnet hosts run `siad`.

I still haven't gotten around to implement sector streaming for uploads, but once I do, I expect we'll see significant gains there as well.